refactor(ux): extract FlowLayout to own file, trim CaptureWindowView to 300 lines
CI file size check requires all Swift files ≤ 300 lines.
FlowLayout is a general-purpose layout used by both CaptureWindowView
and SubredditRow, so it makes sense as its own file anyway.
